test(auth): Added base class authenticationProvider tests (#925)
